+++ crates/icp-cli/src/main.rs
@@ -314,6 +314,12 @@ async fn main() -> Result<(), Error> {
 
         // Network
         Command::Network(cmd) => match cmd {
+            commands::network::Command::Info(args) => {
+                commands::network::info::exec(&ctx, &args)
+                    .instrument(trace_span)
+                    .await?
+            }
+
             commands::network::Command::List(args) => {
                 commands::network::list::exec(&ctx, &args)
                     .instrument(trace_span)